Implement Downsampling in Rust Computer Vision (#47)
* feat: implement downsampling and debug bazel linker issues

* chore: un-exclude crates that were never gonna be included anyways

* address feedback

* address changes

* bruh

* propagate errors, type conversions, no frame clone

* remove intensity normalization, run formatter, rename timestamp to be more specific

+++ backend/src/camera/stream/capture.rs
@@ -136,8 +136,8 @@ pub fn run_capture_thread(
                 if !data.is_empty() {
                     let frame = FrameData {
                         pixels: data,
-                        width: resolution.dimensions().0 as u32,
-                        height: resolution.dimensions().1 as u32,
+                        width: resolution.dimensions().0,
+                        height: resolution.dimensions().1,
                         received_at_ns,
                     };
